equiv: tier-2 observable call trace, bswap, atomics; fix getsockopt signedness
Side-effecting helpers (map_update/delete/push/pop/peek, perf_event_output,
ringbuf_output, get_stackid, trace_printk) append events to a shared trace
region compared as an observable — same call sequence, same args, key/value
sizes from each object's .maps BTF. Errno returns are shared per-call-index
oracles (signext32). Pure env reads keyed by their question: get_current_comm
oracle(size,k), skb_load_bytes over a shared packet array with oracle(off,len).
get/set_retval modeled as a 4-byte shared sysret state region. bswap and
atomics (incl. load-acq/store-rel) modeled exactly, sequential semantics.

Sweep: 456 programs EQUIV / 0 INEQUIV / 2 WAIVED (was 366/0/1); 106 objects
fully proved. Negative controls (mutated update flags, key offset) both
INEQUIV via the trace.

True finding #2: cgroup_getset_retval_getsockopt compared optlen>page_size
signed; C's page_size is __u32 (siblings use __s32) so C compares unsigned —
Rust diverged for negative optlen, invisible to test oracle. Fixed, re-proved,
QEMU gate 11/11 OK. Waived: bpf_flow/flow_dissector_4 (LLVM folds
!(data+thoff) to !data; needs NULL skb->data, kernel-impossible).

+++ progs/cgroup_getset_retval_getsockopt.rs
@@ -33,7 +33,7 @@ static mut retval_value: u32 = 0;
 #[no_mangle]
 static mut ctx_retval_value: u32 = 0;
 #[no_mangle]
-static mut page_size: i32 = 0;
+static mut page_size: u32 = 0;
 
 #[link_section = "cgroup/getsockopt"]
 #[no_mangle]
@@ -49,7 +49,7 @@ extern "C" fn get_retval(ctx: *mut bpf_sockopt) -> i32 {
 
     sync_fetch_and_add_u32(unsafe { core::ptr::addr_of_mut!(invocations) }, 1);
 
-    if ctx.optlen > unsafe { page_size } {
+    if ctx.optlen as u32 > unsafe { page_size } {
         ctx.optlen = 0;
     }
 
@@ -66,7 +66,7 @@ extern "C" fn set_eisconn(ctx: *mut bpf_sockopt) -> i32 {
     }
 
     let ctx = unsafe { &mut *ctx };
-    if ctx.optlen > unsafe { page_size } {
+    if ctx.optlen as u32 > unsafe { page_size } {
         ctx.optlen = 0;
     }
 
@@ -81,7 +81,7 @@ extern "C" fn clear_retval(ctx: *mut bpf_sockopt) -> i32 {
     let ctx = unsafe { &mut *ctx };
     ctx.retval = 0;
 
-    if ctx.optlen > unsafe { page_size } {
+    if ctx.optlen as u32 > unsafe { page_size } {
         ctx.optlen = 0;
     }
